ctypes - Combinators for binding to C libraries without writing any C

What is ctypes?

Ctypes is an OCaml library to interface with existing C libraries, without writing any C "glue code".

Tutorial

Tutorial

How-to guides

How to use ctypes in a dune project?

How to bind constants?

Reference

API documentation

Explanation

How stub generation works

Dynamic linking and dynamic loading, or: "Why can't ctypes find my symbols?"

How does ctypes constant extraction work?